Fix Tipo dato Radio al traer de BD

+++ REApp/REApp/Forms/SolicitudesAnalisis.aspx.cs
@@ -712,7 +712,7 @@ namespace REApp.Forms
                             else
                             {
                                 puntoGeografico.EsPoligono = false;
-                                puntoGeografico.Radio = dtPuntosGeograficos.Rows[j]["Radio"].ToString().ToInt();
+                                puntoGeografico.Radio = dtPuntosGeograficos.Rows[j]["Radio"].ToString().ToDouble();
                                 id++;
                             }
                             puntoGeografico.eliminarBD = false;
